tech(api): move getRequestId from monitoring tools to execution context manager
